backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSockets
websockets==12.0

# CORS
python-cors==1.0.0

# Environment variables
python-dotenv==1.0.0

# Date and time
python-dateutil==2.8.2

# UUID
uuid==1.30

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
